Tests/LibWeb/Text/expected/scroll-left-and-top-on-colgroup.txt

scroll left = 0
scroll top = 0
